Make synthVoice return int instead of short, gives a nice speed up also add a comment
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@14988 a1c6a512-1295-4272-9138-f99709370657

+++ apps/plugins/midi/synth.c
@@ -255,8 +255,8 @@ inline void stopVoice(struct SynthObject * so)
     so->decay = 0;
 }
 
-signed short int synthVoice(struct SynthObject * so) ICODE_ATTR;
-signed short int synthVoice(struct SynthObject * so)
+int synthVoice(struct SynthObject * so) ICODE_ATTR;
+int synthVoice(struct SynthObject * so)
 {
     struct GWaveform * wf;
     register int s;
